checker: move ComputeFingerprint out of formatters
The fingerprint is a change-identity hash (sha256 of id:operation:path:args),
not an output concern, but living in formatters forced callers to import the
rendering package just to compute it. Move it (and serializeArgs) to
checker/fingerprint.go, next to the Change it identifies. Callers switch to
checker.ComputeFingerprint; no behavior change.
