+++ validate/validate.go
@@ -88,8 +88,8 @@ func flattenKinErrors(source string, err error) formatters.Findings {
 	// Fingerprint last so it hashes over the populated fields. validate and
 	// changelog share ComputeFingerprint so a downstream tool can match
 	// findings across spec versions; the args carry the per-finding
-	// disambiguator (formatters/changes.go).
-	f.Fingerprint = formatters.ComputeFingerprint(f.Id, f.Operation, f.Path, argsForKinError(err))
+	// disambiguator (checker/fingerprint.go).
+	f.Fingerprint = checker.ComputeFingerprint(f.Id, f.Operation, f.Path, argsForKinError(err))
 	return formatters.Findings{f}
 }
